fix:修改删除角色有使用的问题

+++ study_zero_module/study_zero_system/src/main/java/com/demo/domain/service/impl/SysRoleServiceImpl.java
@@ -7,7 +7,9 @@ import com.demo.domain.entry.dto.SysRoleDto;
 import com.demo.domain.entry.po.SysRolePo;
 import com.demo.domain.entry.po.SysUserRolePo;
 import com.demo.domain.mapper.SysRoleMapper;
+import com.demo.domain.mapper.SysUserRoleMapper;
 import com.demo.domain.service.SysRoleService;
+import com.demo.enums.ZeroResultEnum;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -24,6 +26,9 @@ public class SysRoleServiceImpl implements SysRoleService{
 
     @Autowired
     private SysRoleMapper sysRoleMapper;
+
+    @Autowired
+    private SysUserRoleMapper sysUserRoleMapper;
 
     @Override
     public List<SysRolePo> queryRoleList(SysRoleDto dto) {
@@ -51,6 +56,13 @@ public class SysRoleServiceImpl implements SysRoleService{
 
     @Override
     public Boolean delRole(SysRolePo sysRolePo) {
+
+        LambdaQueryWrapper<SysUserRolePo> wrapper = new QueryWrapper<SysUserRolePo>().lambda()
+                .eq(SysUserRolePo::getRoleId, sysRolePo.getId());
+        List<SysUserRolePo> sysUserRolePos = sysUserRoleMapper.selectList(wrapper);
+        if(sysUserRolePos.size()>0){
+            throw new RuntimeException(ZeroResultEnum.SYSTEM_ROLE_IS_USER.toString());
+        }
         return SqlHelper.retBool(sysRoleMapper.deleteById(sysRolePo.getId()));
     }
 }
